fix some broken tests
